+++ app/src/main/java/com/example/playlistmaker/domain/db/FavoriteRepository.kt
@@ -8,5 +8,5 @@ interface FavoriteRepository {
     fun addTrack(track:Track)
     fun deleteTrack(track: Track)
     fun getFavorite() : Flow<List<Track>>
-    fun checkTreckInFavorite(id: Long): Flow<Boolean>
+    fun checkTreckInFavorite(id: Int): Flow<Boolean>
 }
